Sort src names, which influence symbol order in executable.
This had to be changed because in recent GNU make releases,
"wildcard" no longer sorts file names.

This is useful if you'd like the run address of your CPC binary equal to
the load address.  For example, name the file containing main()
like "00_main.c" so that its symbol is first in the CPC binary.
If your program uses initialized global variables, it is taken care of
automatically if you use CDTC's crt0 because makefiles always put it first.